修改了软件的更新逻辑

+++ AutoUpdate/UpdateAssistant/Program.cs
@@ -17,17 +17,21 @@ namespace UpdateAssistant
                 return;
             string install_ini = args[0];
             string package_name = args[1];
-            if (!File.Exists(install_ini) || !File.Exists(package_name))
+            string package_zip_name = package_name + ".zip";
+            ProjectFile.UpdateMethod update_method = (ProjectFile.UpdateMethod)Enum.Parse(typeof(ProjectFile.UpdateMethod), args[2]);
+            Console.WriteLine(update_method);
+            Console.Read();
+            return;
+            if (!File.Exists(install_ini) || !File.Exists(package_zip_name))
             {
                 if (File.Exists(install_ini))
                     File.Delete(install_ini);
-                if (File.Exists(package_name))
-                    File.Delete(package_name);
+                if (File.Exists(package_zip_name))
+                    File.Delete(package_zip_name);
                 return;
             }
-            if (package_name.IndexOf(".zip") == -1)
-                package_name = package_name + ".zip";
 
+            Console.Read();
             // 主程序退出后进行替换
             bool flag = true;
             while (flag)
@@ -41,12 +45,22 @@ namespace UpdateAssistant
                     }   
                 }
             }
-            
+
+            if (File.Exists(package_zip_name))
+            {
+                Console.WriteLine(package_zip_name);
+                Console.Read();
+                // 解压程序包
+                ZipFile.ExtractToDirectory(package_zip_name, AppDomain.CurrentDomain.BaseDirectory + "\\" + package_name);
+                // 解压完成后删除程序包
+                File.Delete(package_zip_name);
+            }
+
             DirectoryInfo dic = new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory);
             FileInfo[] files = dic.GetFiles();
             foreach (var item in files)
             {
-                if (Path.GetExtension(item.Name) == ".ini" || item.Name == "UpdateAssistant.exe" || item.Name == package_name)
+                if (Path.GetExtension(item.Name) == ".ini" || item.Name == "UpdateAssistant.exe" || item.Name == package_zip_name)
                     continue;
                 // 删除旧的文件
                 File.Delete(item.Name);
@@ -54,17 +68,11 @@ namespace UpdateAssistant
             DirectoryInfo[] dirs = dic.GetDirectories();
             foreach (var item in dirs)
             {
+                if (item.Name == package_name)
+                    continue;
                 Directory.Delete(item.Name, true);
             }
             
-            if (File.Exists(package_name))
-            {
-                // 解压程序包
-                ZipFile.ExtractToDirectory(package_name, AppDomain.CurrentDomain.BaseDirectory);
-                // 解压完成后删除程序包
-                File.Delete(package_name);
-            }
-
             if (File.Exists("version.ini"))
                 File.Delete("version.ini");
             FileInfo file = new FileInfo(install_ini);
